[update] add index file for TextInput
